index.js to server.js refactor up and running
